pyproject.toml
[project]
name = "beever-atlas"
version = "0.1.0"
description = "Wiki-first RAG system with dual semantic + graph memory"
readme = { text = "", content-type = "text/markdown" }
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.7.0",
    "google-adk>=1.0.0",
    "litellm>=1.75.5",
    "weaviate-client>=4.10.0",
    "neo4j>=5.26.0",
    "pymongo>=4.10.0",
    "motor>=3.6.0",
    "redis>=5.2.0",
    "httpx>=0.28.0",
    "sse-starlette>=2.1.0",
    "pypdf>=4.0.0",
    "cryptography>=42.0.0",
    "jellyfish>=1.0.0",
    "apscheduler>=4.0.0a5",
    "croniter>=1.3.0",
]

[project.optional-dependencies]
nebula = [
    "nebula3-python>=3.4.0",
]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "ruff>=0.8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/beever_atlas"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.ruff]
target-version = "py311"
line-length = 100
